expand the variable sizes

+++ Template/src/application/nodes/NodeConfig.h
@@ -63,14 +63,14 @@ typedef struct
 typedef struct
 {
     const uint8_t id;
-    const uint8_t points;
+    const float points;
     const float   percent;
     const float time; /*[s]*/
     const uint16_t x; /*!<[mm]*/
     const uint16_t y; /*!<[mm]*/
     const uint8_t pool_id;
     const uint16_t angle; /*10 bit, 1 bit =ca. 1°*/
-    uint8_t node_tries; /*!< how many times the node was the current node*/
+    uint16_t node_tries; /*!< how many times the node was the current node*/
     node_state_t node_state;
 }node_param_t;
